requirements.txt
torch>=1.8.0
torchvision>=0.9.0
dlib>=19.21.0
opencv-python>=4.5.1
numpy>=1.19.5
tqdm>=4.61.0
Pillow>=8.2.0
